Правки типизации (#5)
* ci(bxbw): Настройка GHA

* feat(theme): перенос функцонала Боксберри в отдельный класс

* feat(bxbw): дополнительные фильтры

Таск https://hyggeland.bitrix24.ru/company/personal/user/13/tasks/task/view/1147/

* refactor(bxbw): поддержка php 8.1 (начало)

* ver(bxbw): readme & changelog

* feat(bxbw): подстановка города если включен плагин сдека, для получения первого значения цены

* fix(bxbw): вынесено в тему, сделано через хук woocommerce_cart_shipping_packages

* fix(bxbw): изменение типа метода

* fix(bxbw): проверка на наличие объекта при вызове состояния метода доставки

* feat(bxbw): крон для обновления заказов при получении статуса Выдано из боксбери

* ver(bxbw): readme & changelog

* fix(theme): типизация размеров

+++ classes/WC_Boxberry_Parent_Method.php
@@ -370,9 +370,9 @@ class WC_Boxberry_Parent_Method extends WC_Shipping_Method {
 			$height = $depth = $width = 0;
 
 			if ( ! is_null( $currentProduct ) ) {
-				$height = (int) round( $currentProduct->get_height() * $dimensionC );
-				$depth  = (int) round( $currentProduct->get_length() * $dimensionC );
-				$width  = (int) round( $currentProduct->get_width() * $dimensionC );
+				$height = (int) round( (int) $currentProduct->get_height() * (int) $dimensionC );
+				$depth  = (int) round( (int) $currentProduct->get_length() * (int) $dimensionC );
+				$width  = (int) round( (int) $currentProduct->get_width() * (int) $dimensionC );
 			}
 
 			$client = $this->get_client();
